Remove dangerous triggers in MulInternals.dfy
This file uses `/noNLarith`, so it's not completely wrong to have `+` and `-` in
triggers.  However, it's particularly dangerous, because triggers on a
quantifier that's directly in the body of a predicate are immediately visible to
the whole Z3 process, so anyone who includes this file without `/noNLarith` will
run into trouble.

In particular, these triggers are the source of
https://github.com/dafny-lang/dafny/issues/1861.